Fix for quotation marks in path elements
Adding quotation marks (") around path elements that contains spaces is
not required (the semicolon is used as separator between path elements),
but is supported by command prompt and environment configuration in
control panel. This change will prevent the app from throwing an
"Invalid characters in path" exception when one of PATH's elements is
wrapped in double quotes.